Removed RDS support from taiga
This was motivated by wanting to remove R as a dependency
