+++ .pre-commit-config.yaml
@@ -1,8 +1,8 @@
 repos:
-  - repo: https://github.com/psf/black
-    rev: "19.3b0"
-    hooks:
-      - id: black
+  #  - repo: https://github.com/psf/black
+  #    rev: "19.3b0"
+  #    hooks:
+  #      - id: black
   - repo: https://github.com/pre-commit/mirrors-prettier
     rev: "v2.2.1"
     hooks:
