+++ taiga2/conv/columnar.py
@@ -619,57 +619,3 @@ def _convert_tabular_to_csv(
 
         if output_file is not None:
             output.close()
-
-
-from typing import Callable
-
-
-def columnar_to_rds(
-    progress,
-    input_file,
-    temp_file_generator: Callable[[], str],
-    encoding="iso-8859-1",
-    max_rows=None,
-    max_bytes=50 * 1024 * 1024,
-):
-    # two step conversion: First convert to csvs, then for each use R to load CSV and write Rdata file.  Not clear that we can do better
-    # at the moment
-    csv_files = convert_tabular_to_multiple_csvs(
-        input_file,
-        temp_file_generator,
-        ",",
-        encoding="utf-8",
-        max_rows=max_rows,
-        max_bytes=max_bytes,
-    )
-    assert encoding in ["utf-8", "iso-8859-1"]
-    if encoding == "utf-8":
-        r_encoding = "UTF-8"
-    else:
-        r_encoding = "Latin-1"
-    rds_files = []
-    for csv_file in csv_files:
-        destination_file = temp_file_generator()
-
-        script = """
-            data <- read.table({csv_file}, sep=',', head=T, as.is=T, check.names=F, quote='\"', comment.char='', encoding={encoding});
-            saveRDS(data, file={dest})
-            """.format(
-            csv_file=r_escape_str(csv_file),
-            dest=r_escape_str(destination_file),
-            encoding=r_escape_str(r_encoding),
-        )
-
-        handle = subprocess.Popen(
-            ["R", "--vanilla"],
-            stdin=subprocess.PIPE,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-
-        stdout, stderr = handle.communicate(script.encode("utf8"))
-        if handle.returncode != 0:
-            raise Exception("R process failed: %s\n%s" % (stdout, stderr))
-
-        rds_files.append(destination_file)
-    return rds_files
