+++ taiga2/tasks.py
@@ -281,8 +281,6 @@ def get_converter(src_format, dst_format):
     if is_hdf5:
         if dst_format == conversion.CSV_FORMAT:
             return conversion.hdf5_to_csv
-        elif dst_format == conversion.RDS_FORMAT:
-            return conversion.hdf5_to_rds
         elif dst_format == conversion.TSV_FORMAT:
             return conversion.hdf5_to_tsv
         elif dst_format == conversion.GCT_FORMAT:
@@ -292,8 +290,6 @@ def get_converter(src_format, dst_format):
             return conversion.columnar_to_csv
         elif dst_format == conversion.TSV_FORMAT:
             return conversion.columnar_to_tsv
-        elif dst_format == conversion.RDS_FORMAT:
-            return conversion.columnar_to_rds
 
     raise Exception("No conversion for {} to {}".format(src_format, dst_format))
