+++ taiga2/tests/blocked_conv_test.py
@@ -5,7 +5,6 @@ from taiga2.conv import (
     csv_to_columnar,
     columnar_to_csv,
     hdf5_to_gct,
-    columnar_to_rds,
 )
 import os
 import sys
@@ -51,30 +50,3 @@ def test_csv_to_columnar(tmpdir):
 
     assert open(tall_table_file_path, "rt").read() == open(dest_csv, "rt").read()
 
-
-def test_large_columnar_to_rds(tmpdir):
-    import csv
-
-    src_csv = str(tmpdir.join("source.csv"))
-    dst_columnar = str(tmpdir.join("source.columnar"))
-
-    # make sizable source file
-    with open(src_csv, "wt") as fd:
-        w = csv.writer(fd)
-        w.writerow(["X" + str(i) for i in range(100)])
-        for i in range(500):
-            w.writerow(["V"] * 100)
-
-    filename_count = [0]
-
-    def temp_file_generator():
-        f = str(tmpdir.join("t" + str(filename_count[0])))
-        filename_count[0] += 1
-        return f
-
-    csv_to_columnar(StubProgress(), src_csv, dst_columnar)
-    # run, with max bytes set to approximate that we should write out 4 files
-    files = columnar_to_rds(
-        StubProgress(), dst_columnar, temp_file_generator, max_bytes=200 * 500 / 3
-    )
-    assert len(files) == 4
