+++ taiga2/tests/imp_conv_test.py
@@ -2,7 +2,6 @@ import pytest
 import os
 
 import csv
-from taiga2.tests.exp_conv_test import ProgressStub
 
 from sqlalchemy.orm import Session
 
@@ -14,10 +13,17 @@ from taiga2.conv.imp import _get_csv_dims
 from taiga2.conv import (
     csv_to_columnar,
     columnar_to_csv,
-    columnar_to_rds,
     csv_to_hdf5,
     hdf5_to_csv,
 )
+
+
+class ProgressStub:
+    def progress(self, *args, **kwargs):
+        print("progress", args, kwargs)
+
+    def failed(self, *args, **kwargs):
+        print("failed", args, kwargs)
 
 test_files_folder_path = "taiga2/tests/test_files"
 
@@ -131,7 +137,6 @@ def test_get_large_csv_dims_wrong_delimeter(tmpdir):
 def test_non_utf8(tmpdir):
     dest = str(tmpdir.join("dest.columnar"))
     final = str(tmpdir.join("final.csv"))
-    rds_dest = str(tmpdir.join("final.rds"))
 
     csv_to_columnar(None, nonutf8_file_path, dest)
     columnar_to_csv(None, dest, lambda: final)
@@ -146,8 +151,6 @@ def test_non_utf8(tmpdir):
         assert row2["row"] == "2"
         assert row2["value"] == "R"
 
-    # lastly, make sure we don't get an exception when converting to rds because R has its own ideas about encoding
-    columnar_to_rds(None, dest, lambda: rds_dest)
 
 
 def test_matrix_with_full_header_import(tmpdir):
